video Controller is javascript function add

+++ controllers/videoController.js
@@ -48,15 +48,14 @@ export const postUpload = async(req, res) => {
         body: { title, description }, 
         file: { path }, 
     } = req;
-    console.log(path);
     const newVideo = await Video.create({
         fileUrl: path,
         title,
         description,
+        creator: req.user.id
     });
-
-    console.log(newVideo);
-    // To Do : Upload and save video
+    req.user.videos.push(newVideo.id);
+    req.user.save();
     res.redirect(routes.videoDetail(newVideo.id));
 };
 export const videoDetail = async(req, res) => {
@@ -64,7 +63,7 @@ export const videoDetail = async(req, res) => {
         params: { id },
     } = req;
     try{
-        const video = await Video.findById(id);
+        const video = await Video.findById(id).populate("creator");
         res.render('videoDetail', { pageTitle: `${video.title}`, video }); // video는 video: video 와 같다.
     }catch(error) {
         res.redirect(routes.home);
@@ -78,7 +77,11 @@ export const getEditVideo = async(req, res) => {
 
     try{
         const video = await Video.findById(id);
-        res.render('editVideo', { pageTitle: `Edit ${video.title}`, video });
+        if(video.creator !== req.user.id){
+            throw Error();
+        }else{
+            res.render('editVideo', { pageTitle: `Edit ${video.title}`, video });
+        }
     }catch(error) {
         console.log(error);
         res.redirect(routes.home);
@@ -107,7 +110,12 @@ export const deleteVideo = async(req, res) => {
         params: { id },
     } = req;
     try{
-        await Video.findOneAndDelete({ _id: id });
+        const video = await Video.findById(id);
+        if(video.creator !== req.user.id){
+            throw Error();
+        }else{
+            await Video.findOneAndDelete({ _id: id });
+        }
     }catch(error) {
         console.log(error);
     }
